V0.7gamma
The organisms now only approach food if they are hungry. Their enzyme effiency reduces till the point they are no longer efficient. Made the code modular so as to aid the functioning of the next parts of the project.

+++ NSS/Food.cpp
@@ -6,7 +6,7 @@ Food::Food()
     //ctor
 }
 
-void Food::updateAmount(double x)
+void Food::updateAmount(float x)
 {
     amount = x;
 }
